+++ impl/src/test/groovy/io/jsonwebtoken/JwtsTest.groovy
@@ -38,32 +38,25 @@ import java.security.PrivateKey
 import java.security.PublicKey
 import java.security.interfaces.ECPublicKey
 import java.security.interfaces.RSAPublicKey
+import java.time.Instant
+import java.time.temporal.ChronoUnit
 
 import static org.junit.Assert.*
 
 class JwtsTest {
 
-    private static Date dateWithOnlySecondPrecision(long millis) {
-        long seconds = (millis / 1000) as long
-        long secondOnlyPrecisionMillis = seconds * 1000
-        return new Date(secondOnlyPrecisionMillis)
+
+    private static Instant instantWithOnlySecondPrecision(Instant instant) {
+        return instant.truncatedTo(ChronoUnit.SECONDS)
     }
 
-    private static Date now() {
-        Date date = dateWithOnlySecondPrecision(System.currentTimeMillis())
-        return date
+    private static long later() {
+        def instant = laterInstant(10000L)
+        return instant.getEpochSecond();
     }
 
-    private static int later() {
-        def date = laterDate(10000)
-        def seconds = date.getTime() / 1000
-        return seconds as int
-    }
-
-    private static Date laterDate(int seconds) {
-        def millis = seconds * 1000L
-        def time = System.currentTimeMillis() + millis
-        return dateWithOnlySecondPrecision(time)
+    private static Instant laterInstant(long seconds) {
+        return instantWithOnlySecondPrecision(Instant.now().plusSeconds(seconds))
     }
 
     protected static String base64Url(String s) {
@@ -151,7 +144,7 @@ class JwtsTest {
         } catch (MalformedJwtException e) {
             String expected = 'Invalid claims: Invalid JWT Claims \'exp\' (Expiration Time) value: -42-. ' +
                     'String value is not a JWT NumericDate, nor is it ISO-8601-formatted. All heuristics exhausted. ' +
-                    'Cause: Unparseable date: "-42-"'
+                    'Cause: Text \'-42-\' could not be parsed at index 1'
             assertEquals expected, e.getMessage()
         }
     }
@@ -407,11 +400,11 @@ class JwtsTest {
 
     @Test
     void testConvenienceExpiration() {
-        Date then = laterDate(10000)
+        Instant then = laterInstant(10000)
         String compact = Jwts.builder().setExpiration(then).compact()
         Claims claims = Jwts.parser().unsecured().build().parse(compact).payload as Claims
-        def claimedDate = claims.getExpiration()
-        assertEquals then, claimedDate
+        def claimedInstant = claims.getExpiration()
+        assertEquals then, claimedInstant
 
         compact = Jwts.builder().setIssuer("Me")
                 .setExpiration(then) //set it
@@ -424,11 +417,11 @@ class JwtsTest {
 
     @Test
     void testConvenienceNotBefore() {
-        Date now = now() //jwt exp only supports *seconds* since epoch:
+        Instant now = instantWithOnlySecondPrecision(Instant.now()) //jwt exp only supports *seconds* since epoch:
         String compact = Jwts.builder().setNotBefore(now).compact()
         Claims claims = Jwts.parser().unsecured().build().parse(compact).payload as Claims
-        def claimedDate = claims.getNotBefore()
-        assertEquals now, claimedDate
+        def claimedInstant = claims.getNotBefore()
+        assertEquals now, claimedInstant
 
         compact = Jwts.builder().setIssuer("Me")
                 .setNotBefore(now) //set it
@@ -441,11 +434,11 @@ class JwtsTest {
 
     @Test
     void testConvenienceIssuedAt() {
-        Date now = now() //jwt exp only supports *seconds* since epoch:
+        Instant now = instantWithOnlySecondPrecision(Instant.now()) //jwt exp only supports *seconds* since epoch:
         String compact = Jwts.builder().setIssuedAt(now).compact()
         Claims claims = Jwts.parser().unsecured().build().parse(compact).payload as Claims
-        def claimedDate = claims.getIssuedAt()
-        assertEquals now, claimedDate
+        def claimedInstant = claims.getIssuedAt()
+        assertEquals now, claimedInstant
 
         compact = Jwts.builder().setIssuer("Me")
                 .setIssuedAt(now) //set it